Move calculators for promotions into their own namespace
They are only used in a promotion context but their names seem
generic: they are not. This can lead to confusion and this change
allows us to defined behavior specific for promotions since now
it's clear that they will be only used in this context.

Previous classes have been deprecated.
