+++ api/spec/requests/spree/api/promotion_application_spec.rb
@@ -12,7 +12,7 @@ module Spree::Api
       let!(:order) { create(:order_with_line_items, line_items_count: 1) }
       let!(:promotion) do
         promotion = create(:promotion, name: "10% off", code: "10off")
-        calculator = Spree::Calculator::FlatPercentItemTotal.create(preferred_flat_percent: "10")
+        calculator = Spree::Calculator::Promotion::FlatPercentItemTotal.create(preferred_flat_percent: "10")
         action = Spree::Promotion::Actions::CreateItemAdjustments.create(calculator: calculator)
         promotion.actions << action
         promotion
