+++ backend/spec/features/admin/promotion_adjustments_spec.rb
@@ -43,7 +43,7 @@ describe "Promotion Adjustments", type: :feature, js: true do
       first_action = promotion.actions.first
       expect(first_action.class).to eq(Spree::Promotion::Actions::CreateAdjustment)
       first_action_calculator = first_action.calculator
-      expect(first_action_calculator.class).to eq(Spree::Calculator::FlatRate)
+      expect(first_action_calculator.class).to eq(Spree::Calculator::Promotion::FlatRate)
       expect(first_action_calculator.preferred_amount).to eq(5)
     end
 
@@ -71,7 +71,7 @@ describe "Promotion Adjustments", type: :feature, js: true do
       first_action = promotion.actions.first
       expect(first_action.class).to eq(Spree::Promotion::Actions::CreateAdjustment)
       first_action_calculator = first_action.calculator
-      expect(first_action_calculator.class).to eq(Spree::Calculator::FlatRate)
+      expect(first_action_calculator.class).to eq(Spree::Calculator::Promotion::FlatRate)
       expect(first_action_calculator.preferred_amount).to eq(5)
     end
 
@@ -106,7 +106,7 @@ describe "Promotion Adjustments", type: :feature, js: true do
       first_action = promotion.actions.first
       expect(first_action.class).to eq(Spree::Promotion::Actions::CreateAdjustment)
       first_action_calculator = first_action.calculator
-      expect(first_action_calculator.class).to eq(Spree::Calculator::FlatPercentItemTotal)
+      expect(first_action_calculator.class).to eq(Spree::Calculator::Promotion::FlatPercentItemTotal)
       expect(first_action_calculator.preferred_flat_percent).to eq(10)
     end
 
@@ -142,7 +142,7 @@ describe "Promotion Adjustments", type: :feature, js: true do
       first_action = promotion.actions.first
       expect(first_action.class).to eq(Spree::Promotion::Actions::CreateItemAdjustments)
       first_action_calculator = first_action.calculator
-      expect(first_action_calculator.class).to eq(Spree::Calculator::PercentOnLineItem)
+      expect(first_action_calculator.class).to eq(Spree::Calculator::Promotion::PercentOnLineItem)
       expect(first_action_calculator.preferred_percent).to eq(10)
     end
 
@@ -236,7 +236,7 @@ describe "Promotion Adjustments", type: :feature, js: true do
 
       first_action = promotion.actions.first
       expect(first_action.class).to eq(Spree::Promotion::Actions::CreateAdjustment)
-      expect(first_action.calculator.class).to eq(Spree::Calculator::FlatRate)
+      expect(first_action.calculator.class).to eq(Spree::Calculator::Promotion::FlatRate)
       expect(first_action.calculator.preferred_amount).to eq(5)
     end
