+++ core/app/models/spree/calculator/flat_rate.rb
@@ -3,16 +3,12 @@
 require_dependency 'spree/calculator'
 
 module Spree
-  class Calculator::FlatRate < Calculator
-    preference :amount, :decimal, default: 0
-    preference :currency, :string, default: ->{ Spree::Config[:currency] }
-
-    def compute(object = nil)
-      if object && preferred_currency.casecmp(object.currency).zero?
-        preferred_amount
-      else
-        0
-      end
-    end
+  class Calculator
+    FlatRate =
+      ActiveSupport::Deprecation::DeprecatedConstantProxy.new(
+        'Spree::Calculator::FlatRate',
+        'Spree::Calculator::Promotion::FlatRate',
+        Spree::Deprecation
+      )
   end
 end
