+++ core/spec/models/spree/calculator/promotion/percent_per_item_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+require 'shared_examples/calculator_shared_examples'
+
+RSpec.describe Spree::Calculator::Promotion::PercentPerItem, type: :model do
+  it_behaves_like 'a calculator with a description'
+end
