+++ core/spec/models/spree/promotion/actions/create_adjustment_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Spree::Promotion::Actions::CreateAdjustment, type: :model do
   # From promotion spec:
   context "#perform" do
     before do
-      action.calculator = Spree::Calculator::FlatRate.new(preferred_amount: 10)
+      action.calculator = Spree::Calculator::Promotion::FlatRate.new(preferred_amount: 10)
       promotion.promotion_actions = [action]
       allow(action).to receive_messages(promotion: promotion)
     end
@@ -82,7 +82,7 @@ RSpec.describe Spree::Promotion::Actions::CreateAdjustment, type: :model do
 
   shared_examples "destroying adjustments from incomplete orders" do
     before(:each) do
-      action.calculator = Spree::Calculator::FlatRate.new(preferred_amount: 10)
+      action.calculator = Spree::Calculator::Promotion::FlatRate.new(preferred_amount: 10)
       promotion.promotion_actions = [action]
     end
